chore: moved object info panel provider to explore folder
