+++ packages/javascript/bh-shared-ui/src/views/Explore/providers/ObjectInfoPanelProvider/index.ts
